move map_files_to_node to utils/nodes.py
Makes it easier to test!
